+++ src/backend/resolvers/queries/people/DEV_pathwaysInfluencers.js
@@ -1,0 +1,5 @@
+module.exports = (parent, args, { pulseDevDb }) =>
+  pulseDevDb
+    .collection('pathwaysInfluencers')
+    .find({ personId: { $ne: null } })
+    .toArray()