+++ src/backend/resolvers/queries/people/index.js
@@ -1,0 +1,9 @@
+const people = require('./people')
+const DEV_pathwaysInfluencers = require('./DEV_pathwaysInfluencers')
+const DEV_providerInfluencers = require('./DEV_providerInfluencers')
+
+module.exports = {
+  DEV_pathwaysInfluencers,
+  DEV_providerInfluencers,
+  people,
+}