+++ src/backend/resolvers/queries/people/people.js
@@ -1,0 +1,2 @@
+module.exports = (parent, args, { pulseCoreDb }) =>
+  pulseCoreDb.collection('people').find().toArray()