+++ src/frontend/Orion/Organizations/Obm/TemplateTable/custom-sort-types/index.js
@@ -1,9 +1,11 @@
 import sortText from './sortText'
 import sortFloat from './sortFloat'
+import sortBool from './sortBool'
 
 const customSortTypes = {
   text: sortText,
   float: sortFloat,
+  bool: sortBool,
 }
 
 export default customSortTypes