+++ src/frontend/components/BusinessObjectModal/OncologyBenefitManagerModal/relational-widgets/ObmInfluencersWidget.js
@@ -11,13 +11,15 @@ import {
   GET_PEOPLE,
   GET_OBM_AND_PERSON_CONNECTIONS,
   GET_INFLUENCER_TEMPLATE_OBMS,
-} from '../../../../api/queries'
+} from 'frontend/api/queries'
 
-import { CONNECT_OBM_AND_PERSON } from '../../../../api/mutations'
+import { CONNECT_OBM_AND_PERSON } from 'frontend/api/mutations'
 
-import { customSelectStyles } from '../../../../components/customSelectStyles'
-import Button from '../../../../components/Button'
-import Color from '../../../../utils/color'
+import useObmPersonConnections from 'frontend/Orion/Organizations/Obm/useObmPersonConnections'
+
+import { customSelectStyles } from 'frontend/components/customSelectStyles'
+import Button from 'frontend/components/Button'
+import Color from 'frontend/utils/color'
 
 import {
   RelationalRow,
@@ -35,12 +37,9 @@ import {
 const ObmInfluencersWidget = ({ entity }) => {
   const { data: peopleData, loading: peopleLoading } = useQuery(GET_PEOPLE)
 
-  const { data: connectionsData, loading: connectionsLoading } = useQuery(
-    GET_OBM_AND_PERSON_CONNECTIONS,
-    {
-      variables: { obmId: entity._id },
-    }
-  )
+  const { data: connectionsData, loading: connectionsLoading } = useObmPersonConnections({
+    obmId: entity._id,
+  })
 
   const [stagedConnections, stageConnections] = useState([])
 
@@ -53,7 +52,6 @@ const ObmInfluencersWidget = ({ entity }) => {
     refetchQueries: [
       {
         query: GET_OBM_AND_PERSON_CONNECTIONS,
-        variables: { obmId: entity._id },
       },
       {
         query: GET_INFLUENCER_TEMPLATE_OBMS,
@@ -64,14 +62,8 @@ const ObmInfluencersWidget = ({ entity }) => {
 
   useEffect(() => {
     if (!peopleLoading && !connectionsLoading) {
-      // ! HOTFIX: make sure there are no connections in the cache for removed people
-      const peopleById = _.keyBy(peopleData.people, '_id')
-      const validConnections = connectionsData.obmAndPersonConnections.filter(
-        (connection) => peopleById[connection.personId]
-      )
-
       // clean data of __typename and anything else
-      const initialConnections = validConnections.map(({ _id, personId, obmId, position }) => ({
+      const initialConnections = connectionsData.map(({ _id, personId, obmId, position }) => ({
         _id,
         personId,
         obmId,